(No Logic) Move existing adjust-payment-dialog.component into new adjust-payment-dialog component folder
